revert to old syntax pattern
